Fixed timing loop
Fixed by removing it. PotSensor data is only sent now when requested from the roomPI as oppsoed to every 10s. This was causing backlog on the serial and causing the arduino to freeze. This freezing also was related to the relay turning off. The freezing has appeared to stop with these current changes